Rewrite boot/debug.c

+++ boot/debug.c
@@ -4,7 +4,7 @@
 // This function is called when an assertion fails and
 // shows the file and line that caused the failure
 
-void error_check(char *file, uint64_t line)
+void error_check(char *file, int line)
 {
 	printk("Assertion Failed [file %s: line %u]", file, line);
 	while (1) { }
